Replace standard IO streams by file IO.
ISO C++11 mandates standard IO streams be thread safe. The side effect
of this is that file IO through shell redirection get lock contention
performance issues. We provide command options for the user to to
specify the input and output files. In this case, we use fstream for
file IO, which doesn't suffer lock contention problems.